+++ SpaceBattle.Lib/Command/StopMoveCommand.cs
