+++ SpaceBattle.Lib/Strategies/IStrategy.cs
